requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# backwards compatibility
pathlib2

#exported from conda env
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
alabaster==0.7.12
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.1
async-generator==1.10
async-timeout==4.0.2
asyncssh==2.8.1
atpublic==2.3
attrs==21.4.0
awscli==1.22.27
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
binaryornot==0.4.4
black==21.12b0
bleach==4.1.0
botocore==1.23.27
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.3
commonmark==0.9.1
configobj==5.0.6
cookiecutter==1.7.3
coverage==6.2
cryptography==36.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.6.0
docopt==0.6.2
docutils==0.15.2
dpath==2.0.5
dulwich==0.20.26
dvc==2.9.3
entrypoints==0.3
flake8==4.0.1
flatten-dict==0.4.2
flit_core==3.6.0
flufl.lock==6.0
fonttools==4.28.5
frozenlist==1.2.0
fsspec==2021.11.1
ftfy==6.0.3
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
grandalf==0.6
httplib2==0.20.2
idna==3.1
imagesize==1.3.0
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.3.3
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
mailchecker==4.1.7
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
multidict==5.2.0
munkres==1.1.4
mypy==0.930
mypy-extensions==0.4.3
nanotime==0.5.2
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numpy==1.21.5
oauth2client==4.1.3
olefile==0.46
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
phonenumbers==8.12.40
pickleshare==0.7.5
Pillow==8.4.0
pip==21.3.1
pipreqs==0.4.11
platformdirs==2.4.1
ply==3.11
poyo==0.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycosat==0.6.3
pycparser==2.21
pydot==1.4.2
PyDrive2==1.10.0
pyflakes==2.4.0
pygit2==1.7.2
Pygments==2.11.1
pygtrie==2.4.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
python-benedict==0.24.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.26.0
rich==10.16.2
rsa==4.7.2
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scikit-learn==1.0.2
scipy==1.7.3
scmrepo==0.0.7
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==60.2.0
shortuuid==1.0.8
shtab==1.5.3
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.3.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
src==0.1.0
statsmodels==0.13.1
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.3
torch==1.10.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
voluptuous==0.12.2
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==3.5.2
xmltodict==0.12.0
yarg==0.1.9
yarl==1.7.2
zc.lockfile==2.0
zipp==3.6.0
